requirements.txt
django
djangorestframework
bpython
